[#9005] ysql: Verify pid corresponds to postgres process before killing leftover postgres processes
Summary:
Currently the tserver kills leftover postgres processes from a previous run if postmaster.pid file is still present.

In the case where there is an unexpected shutdown of postgres/tserver, followed by a reboot of the VM or restart of the k8s pod, the postmater.pid file could have a PID but that PID no longer corresponds to a postgres process. In this case, the code ends up killing a potentially random process.

This revision adds check that the process name is a postgres binary before killing.

Test Plan: Existing test suite.

Reviewers: mihnea, sanketh

Reviewed By: sanketh

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D12046

+++ src/yb/yql/pgwrapper/pg_wrapper.cc
@@ -579,8 +579,24 @@ CHECKED_STATUS PgSupervisor::CleanupOldServerUnlocked() {
       // If process does not exist, system may return "process does not exist" or
       // "operation not permitted" error. Ignore those errors.
       postmaster_pid_file.close();
-      if (kill(postgres_pid, SIGKILL) != 0 && errno != ESRCH && errno != EPERM) {
-        return STATUS(RuntimeError, "Unable to kill", Errno(errno));
+      bool postgres_found = true;
+      string cmdline = "";
+#ifdef __linux__
+      string cmd_filename = "/proc/" + std::to_string(postgres_pid) + "/cmdline";
+      std::ifstream postmaster_cmd_file;
+      postmaster_cmd_file.open(cmd_filename, std::ios_base::in);
+      if (postmaster_cmd_file.good()) {
+        postmaster_cmd_file >> cmdline;
+        postgres_found = cmdline.find("/postgres") != std::string::npos;
+        postmaster_cmd_file.close();
+      }
+#endif
+      if (postgres_found) {
+        if (kill(postgres_pid, SIGKILL) != 0 && errno != ESRCH && errno != EPERM) {
+          return STATUS(RuntimeError, "Unable to kill", Errno(errno));
+        }
+      } else {
+        LOG(WARNING) << "Didn't find postgres in " << cmdline;
       }
     }
     ignore_result(Env::Default()->DeleteFile(postmaster_pid_filename));
